ml-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow==2.16.1  # Usa la misma versión de tu notebook o una compatible
pandas
numpy
Pillow